storage/requirements.txt
connexion
requests
SQLAlchemy
mysql-connector
PyMySQL
PyYAML
connexion[swagger-ui]